+++ emg3d/__init__.py
@@ -14,6 +14,20 @@
 # License for the specific language governing permissions and limitations under
 # the License.
 
+# Import modules
+from emg3d import electrodes
+from emg3d import fields
+from emg3d import io
+from emg3d import maps
+from emg3d import meshes
+from emg3d import models
+from emg3d import optimize
+from emg3d import simulations
+from emg3d import solver
+from emg3d import surveys
+from emg3d import utils
+
+# Import most important functions and classes
 from emg3d.fields import Field, get_source_field, get_magnetic_field
 from emg3d.io import save, load
 from emg3d.meshes import TensorMesh, construct_mesh